+++ crates/hypercolor-core/tests/servo_css_probe_tests.rs
@@ -288,7 +288,7 @@ fn frame_input(frame_number: u64, canvas_width: u32, canvas_height: u32) -> Fram
 
     #[allow(clippy::cast_precision_loss)]
     FrameInput {
-        time_secs: frame_number as f32 * FRAME_DT_SECONDS,
+        time_secs: frame_number as f64 * f64::from(FRAME_DT_SECONDS),
         delta_secs: FRAME_DT_SECONDS,
         frame_number,
         audio: &AUDIO,
